+++ boundbox-library/src/test/java/org/boundbox/processor/BoundBoxProcessorTest.java
@@ -944,6 +944,33 @@ public class BoundBoxProcessorTest {
         assertTrue(listInnerClassInfos.isEmpty());
     }
     
+    @Test
+    public void testProcess_class_is_static_inner_class_that_extends_non_static_inner_class() throws URISyntaxException {
+        // given
+        String[] testSourceFileNames = new String[] { "TestClassIsStaticInnerClassThatExtendsNonStaticInnerClass.java" };
+        CompilationTask task = processAnnotations(testSourceFileNames, boundBoxProcessor);
+
+        // when
+        // Perform the compilation task.
+        task.call();
+
+        // then
+        assertFalse(boundBoxProcessor.getListClassInfo().isEmpty());
+        ClassInfo classInfo = boundBoxProcessor.getListClassInfo().get(0);
+
+        List<MethodInfo> listMethodInfos = classInfo.getListMethodInfos();
+        assertTrue(listMethodInfos.isEmpty());
+        List<FieldInfo> listFieldInfos = classInfo.getListFieldInfos();
+        FakeFieldInfo fakeFieldInfo = new FakeFieldInfo("foo", int.class.getSimpleName());
+        fakeFieldInfo.setStaticField(false);
+        fakeFieldInfo.setInheritanceLevel(1);
+        fakeFieldInfo.setEffectiveInheritanceLevel(0);
+        assertContains(listFieldInfos, fakeFieldInfo);
+
+        List<InnerClassInfo> listInnerClassInfos = classInfo.getListInnerClassInfo();
+        assertTrue(listInnerClassInfos.isEmpty());
+    }
+    
     // ----------------------------------
     // IMPORTS
     // ----------------------------------